% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAdditionalTransportSources.R
\name{calcAdditionalTransportSources}
\alias{calcAdditionalTransportSources}
\title{Provide additional data sources for Transport validation}
\usage{
calcAdditionalTransportSources(subtype, FiveYearAverage = TRUE)
}
\arguments{
\item{subtype}{one of the source parameters}

\item{FiveYearAverage}{switch to enable smoothing by calculating centered 5 years avergage}
}
\description{
Provide additional data sources for Transport validation
}
\author{
Johanna Hoppe
}
