% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEAOutputTransport.R
\name{calcIEAOutputTransport}
\alias{calcIEAOutputTransport}
\title{Calc Input Output}
\usage{
calcIEAOutputTransport()
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes IEA-based model data by use of raw IEA "Energy Balances" data
and a mapping that corresponds to the structure of "products" and "flows" of IEA.
}
\details{
Mapping structure example: IEA product ANTCOAL used for IEA flow TPATFUEL, contributes via REMIND technology
coaltr for generating sesofos from pecoal (REMIND names)
}
\author{
Falk Benke
}
