% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertACEA.R
\name{convertACEA}
\alias{convertACEA}
\title{Convert ACEA vehicles on roads data to iso country.}
\usage{
convertACEA(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert ACEA vehicles on roads data to iso country.
}
\examples{
\dontrun{
a <- readSource("ACEA")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
