% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEUROSTAT.R
\name{convertEUROSTAT}
\alias{convertEUROSTAT}
\title{Convert Eurostat road transportation data to iso country.}
\usage{
convertEUROSTAT(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert Eurostat road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("EUROSTAT")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Johanna Hoppe
}
