% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEUstatisticalPocketbook.R
\name{convertEUstatisticalPocketbook}
\alias{convertEUstatisticalPocketbook}
\title{Convert EU statistical pocket book data to iso country.}
\usage{
convertEUstatisticalPocketbook(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert EU statistical pocket book data to iso country.
}
\examples{
\dontrun{
a <- readSource("EUstatisticalPocketbook")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
