% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJRC.R
\name{convertJRC}
\alias{convertJRC}
\title{Convert Eurostat road transportation data to iso country.}
\usage{
convertJRC(x)
}
\arguments{
\item{x}{a magpie data object}
}
\value{
magpie object
}
\description{
Convert Eurostat road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("JRC")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
