% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTransportPurchasePriceSubsidies.R
\name{convertTransportPurchasePriceSubsidies}
\alias{convertTransportPurchasePriceSubsidies}
\title{Converts transport subsidies data}
\usage{
convertTransportPurchasePriceSubsidies(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing transport subsidies per technology
}
\description{
Converts transport subsidies data
}
\examples{

\dontrun{ a <- convertTransportSubsidies(x)
}

}
\author{
Caroline Cronjaeger
}
