% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCD.R
\name{convertUCD}
\alias{convertUCD}
\title{Convert UCD road transportation data to iso country.}
\usage{
convertUCD(x, subtype)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Convert UCD road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("UCD")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Alois Dirnaichner
}
