% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGCAM.R
\name{correctGCAM}
\alias{correctGCAM}
\title{Correct GCAM road transportation data to iso country.}
\usage{
correctGCAM(magpieobj, subtype)
}
\arguments{
\item{magpieobj}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Correct GCAM road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("GCAM", subtype="histEsDemand")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Alois Dirnaichner
}
