% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readACEA.R
\name{readACEA}
\alias{readACEA}
\title{Read ACEA vehicles on roads data}
\usage{
readACEA(subtype = c("historicalEnergyServiceDemand"))
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read ACEA vehicles on roads data
}
\examples{
\dontrun{
a <- readSource("ACEA")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
