% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanAlternativeFuelsObservatory.R
\name{readEuropeanAlternativeFuelsObservatory}
\alias{readEuropeanAlternativeFuelsObservatory}
\title{Read European Alternative Fuels Observatory data}
\usage{
readEuropeanAlternativeFuelsObservatory(
  subtype = c("historicalVehStockTechShares")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read European Alternative Fuels Observatory data
}
\examples{
\dontrun{
a <- readSource("EEA")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
