% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostatEnergyCountryDataSheets.R
\name{readEurostatEnergyCountryDataSheets}
\alias{readEurostatEnergyCountryDataSheets}
\title{Read Eurostat data.}
\usage{
readEurostatEnergyCountryDataSheets(subtype = c("feDemand", "LDVfleet"))
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read Eurostat data.
}
\examples{
\dontrun{
a <- readSource("EurostatEnergyCountryDataSheets")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
