% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEUstatisticalPocketbook.R
\name{readEUstatisticalPocketbook}
\alias{readEUstatisticalPocketbook}
\title{Read EU statistical pocketbook transport data}
\usage{
readEUstatisticalPocketbook(
  subtype = c("historicalEnergyServiceDemand", "historicalVehStock",
    "historicalVehSales")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read EU statistical pocketbook transport data
}
\examples{
\dontrun{
a <- readSource("EUstatisticalPocketbook")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
