% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCAM.R
\name{readGCAM}
\alias{readGCAM}
\title{Read GCAM road transportation data.}
\usage{
readGCAM(
  subtype = c("feVkmIntensity", "loadFactor", "histESdemand", "speedMotorized",
    "speedNonMotorized", "valueOfTimeMultiplier")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Read GCAM road transportation data.
}
\seealso{
\code{\link{readSource}}
}
\author{
Johnna Hoppe, Alois Dirnaichner
}
