% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJRC.R
\name{readJRC}
\alias{readJRC}
\title{Read EU Joint Research Center data.}
\usage{
readJRC(subtype = c("feDemand", "LDVfleet"))
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read EU Joint Research Center data.
}
\examples{
\dontrun{
a <- readSource("EurostatEnergyCountryDataSheets")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
