% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPSI.R
\name{readPSI}
\alias{readPSI}
\title{Read PSI data.}
\usage{
readPSI(subtype = c("CAPEX", "energyIntensity"))
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Read PSI data.
}
\examples{
\dontrun{
a <- readSource("PSI", subtype = "energyIntensity")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johnna Hoppe
}
