% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMINDinputForTransportStandalone.R
\name{readREMINDinputForTransportStandalone}
\alias{readREMINDinputForTransportStandalone}
\title{Load fuel prices from a REMIND fulldata.gdx in USD/MJ}
\usage{
readREMINDinputForTransportStandalone()
}
\description{
Load fuel prices from a REMIND fulldata.gdx in USD/MJ
}
