% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTRACCS.R
\name{readTRACCS}
\alias{readTRACCS}
\title{Read TRACCS road transportation data.}
\usage{
readTRACCS(
  subtype = c("fuelEnDensity", "roadFuelConsumption", "energyIntensity", "loadFactor",
    "annualMileage", "roadESdemand", "histESdemand", "railFeDemand", "vehPopulation")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read TRACCS road transportation data.
}
\examples{
\dontrun{
a <- readSource("TRACCS")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Alois Dirnaichner
}
