% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTransportPurchasePriceSubsidies.R
\name{readTransportPurchasePriceSubsidies}
\alias{readTransportPurchasePriceSubsidies}
\title{Read transport subsidies data}
\usage{
readTransportPurchasePriceSubsidies()
}
\value{
magpie object of the transport subsidies for BEV, FCEV and PHEV (euros/car) for private and legal entities
}
\description{
Read-in transport subsidies csv files as magclass object
}
\examples{
\dontrun{ a <- readSource(type="TransportSubsidies")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Caroline Cronjaeger
}
