% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUCD.R
\name{readUCD}
\alias{readUCD}
\title{Read UCD road transportation data.}
\usage{
readUCD(
  subtype = c("energyIntensity", "feDemand", "loadFactor", "annualMileage",
    "nonMotorizedDemand", "speed", "CAPEX", "nonFuelOPEX", "CAPEXandNonFuelOPEX",
    "OperatingSubsidies")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read UCD road transportation data.
}
\examples{
\dontrun{
a <- readSource("UCD")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe, Alois Dirnaichner
}
