% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustCAPEXother.R
\name{toolAdjustCAPEXother}
\alias{toolAdjustCAPEXother}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustCAPEXother(dt, ISOcountries, yrs, completeData)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{ISOcountries}{list of iso countries}

\item{yrs}{temporal resolution of EDGE-T model}

\item{completeData}{complete EDGE-T decision tree}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
