% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustEnergyIntensity.R
\name{toolAdjustEnergyIntensity}
\alias{toolAdjustEnergyIntensity}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustEnergyIntensity(
  dt,
  regionTRACCS,
  TrendsEnIntPSI,
  filter,
  ariadneAdjustments = TRUE
)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{regionTRACCS}{iso countries included in TRACCS database}

\item{TrendsEnIntPSI}{Energy intensity trends from PSI}

\item{filter}{list of filters for specific branches in the upper decision tree, containing all associated
univocalNames}

\item{ariadneAdjustments}{optional parameter adjustements according to ARIADNE model intercomparison in 2022}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
