% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustLoadFactor.R
\name{toolAdjustLoadFactor}
\alias{toolAdjustLoadFactor}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustLoadFactor(dt, completeData, TRACCScountries)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{completeData}{complete EDGE-T decision tree}

\item{TRACCScountries}{countries included in TRACCS database}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
