% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustSpeedOfModes.R
\name{toolAdjustSpeedOfModes}
\alias{toolAdjustSpeedOfModes}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustSpeedOfModes(dt, completeData)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{completeData}{complete EDGE-T decision tree}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
