% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregatVehicleTypes.R
\name{toolAggregateVehicleTypes}
\alias{toolAggregateVehicleTypes}
\title{Aggregate vehicle types so that vehicle types that are only available in certain iso countries of a region
are not lost and still appear in the aggregated region}
\usage{
toolAggregateVehicleTypes(...)
}
\arguments{
\item{...}{magpie object to be aggregated}
}
\value{
a magpie object
}
\description{
Aggregate vehicle types so that vehicle types that are only available in certain iso countries of a region
are not lost and still appear in the aggregated region
}
\author{
Johanna Hoppe
}
