% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMergeEnergyIntensity.R
\name{toolMergeEnergyIntensity}
\alias{toolMergeEnergyIntensity}
\title{Merge source data for energy intensity}
\usage{
toolMergeEnergyIntensity(data, filterEntries, countriesTRACCS)
}
\arguments{
\item{data}{source data}

\item{filterEntries}{helper to filter for univocalNames}

\item{countriesTRACCS}{countries included in the TRACCS database}
}
\description{
Merge source data for energy intensity
}
\author{
Johanna Hoppe
}
