% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMergeHistESdemand.R
\name{toolMergeHistESdemand}
\alias{toolMergeHistESdemand}
\title{Merge source data for historical energy service demand}
\usage{
toolMergeHistESdemand(data, filterEntries, countriesTRACCS)
}
\arguments{
\item{data}{source data}

\item{filterEntries}{helper to filter for univocalNames}

\item{countriesTRACCS}{countries included in the TRACCS database}
}
\description{
Merge source data for historical energy service demand
}
\author{
Johanna Hoppe
}
