% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareACEA.R
\name{toolPrepareACEA}
\alias{toolPrepareACEA}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareACEA(x)
}
\arguments{
\item{x}{the input data read via readSource, a magpie object}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories.
}
\author{
Johanna Hoppe
}
