% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/toolPrepareEuropeanAlternativeFuelsObservatory.R
\name{toolPrepareEuropeanAlternativeFuelsObservatory}
\alias{toolPrepareEuropeanAlternativeFuelsObservatory}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareEuropeanAlternativeFuelsObservatory(x)
}
\arguments{
\item{x}{the input data read via readSource, a magpie objec}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories.
}
\author{
Johanna Hoppe
}
