% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareEurostatEnergyCountryDataSheets.R
\name{toolPrepareEurostatEnergyCountryDataSheets}
\alias{toolPrepareEurostatEnergyCountryDataSheets}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareEurostatEnergyCountryDataSheets(x)
}
\arguments{
\item{x}{the input data read via readSource, a magpie object}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories. Apply the full structure of the decision tree.
}
\author{
Johanna Hoppe
}
