% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareJRC.R
\name{toolPrepareJRC}
\alias{toolPrepareJRC}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareJRC(
  x,
  subtype = c("energyServiceDemand", "energyServiceDemandTechnologyLevel")
)
}
\arguments{
\item{x}{the input data read via readSource, a magpie object}

\item{subtype}{One of the possible subtypes}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories. Apply the full structure of the decision tree.
}
\author{
Johanna Hoppe
}
