% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPreparePSI.R
\name{toolPreparePSI}
\alias{toolPreparePSI}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPreparePSI(x)
}
\arguments{
\item{x}{the input data read via readSource, a magpie object}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories. Apply the full logit structure.
}
\author{
Johanna Hoppe
}
