% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareTransportSubsidies.R
\name{toolPrepareTransportSubsidies}
\alias{toolPrepareTransportSubsidies}
\title{Prepare subsidies data}
\usage{
toolPrepareTransportSubsidies(helpers)
}
\arguments{
\item{helpers}{list containg several helpers used throughout the model.
It includes dtTimeRes, a data.table containing the temporal
resolution for different univocalNames}
}
\value{
list of data.tables containing mrremind input data
}
\description{
Prepare subsidies data
}
