% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareUCD.R
\name{toolPrepareUCD}
\alias{toolPrepareUCD}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareUCD(x, subtype)
}
\arguments{
\item{x}{the input data read via readSource, a magpie object}

\item{subtype}{one of the different EDGE-T inputdata subtypes}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories. Apply the full structure of the decision tree.
}
\author{
Johanna Hoppe
}
