% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEdgeTransportSAinputs.R
\name{calcEdgeTransportSAinputs}
\alias{calcEdgeTransportSAinputs}
\title{Provide EDGE-Transport input parameters}
\usage{
calcEdgeTransportSAinputs(subtype, IEAharm = TRUE)
}
\arguments{
\item{subtype}{one of the parameters required for EDGE-T SA}

\item{IEAharm}{switch IEA harmonization of energy intensity on and off}
}
\description{
Provide EDGE-Transport input parameters
}
\author{
Johanna Hoppe
}
