% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostatEnergyCountryDataSheets.R
\name{convertEurostat}
\alias{convertEurostat}
\title{Convert Eurostat road transportation data to iso country.}
\usage{
convertEurostat(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert Eurostat road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("Eurostat")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Johanna Hoppe
}
