% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGCAM.R
\name{convertGCAM}
\alias{convertGCAM}
\title{Convert GCAM road transportation data to iso country.}
\usage{
convertGCAM(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert GCAM road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("GCAM", subtype = "esDemand")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Johanna Hoppe, Alois Dirnaichner
}
