% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGCAM.R
\name{convertGCAM}
\alias{convertGCAM}
\title{Convert GCAM transportation data to iso country level.}
\usage{
convertGCAM(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert GCAM transportation data to iso country level.
}
\examples{
\dontrun{
a <- readSource("GCAM", subtype = "esDemand")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe, Alois Dirnaichner
}
