% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPSI.R
\name{convertPSI}
\alias{convertPSI}
\title{Convert PSI data to iso country.}
\usage{
convertPSI(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert PSI data to iso country.
}
\examples{
\dontrun{
a <- readSource("PSI", subtype = "costs")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe
}
