% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUCD.R
\name{convertUCD}
\alias{convertUCD}
\title{Convert UCD road transportation data to iso country.}
\usage{
convertUCD(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Convert UCD road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("UCD")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe, Alois Dirnaichner
}
