% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGCAM.R
\name{correctGCAM}
\alias{correctGCAM}
\title{Correct GCAM road transportation data to iso country.}
\usage{
correctGCAM(x, subtype)
}
\arguments{
\item{x}{a magpie data object}

\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Correct GCAM road transportation data to iso country.
}
\examples{
\dontrun{
a <- readSource("GCAM", subtype="histESdemand")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johanna Hoppe, Alois Dirnaichner
}
