% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtransport-package.R
\docType{package}
\name{mrtransport-package}
\alias{mrtransport}
\alias{mrtransport-package}
\title{mrtransport: Input data generation for the EDGE-Transport model}
\description{
The mrtransport package contains data preprocessing for the EDGE-Transport model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrtransport}
}

}
\author{
\strong{Maintainer}: Johanna Hoppe \email{johanna.hoppe@pik-potsdam.de} (\href{https://orcid.org/0009-0004-6753-5090}{ORCID})

Authors:
\itemize{
  \item Alois Dirnaichner
}

}
