% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEUROSTAT.R
\name{readEUROSTAT}
\alias{readEUROSTAT}
\title{Read UCD road transportation data.}
\usage{
readEUROSTAT(subtype = c("feDemand", "LDVfleet"))
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magclass object
}
\description{
Read UCD road transportation data.
}
\examples{
\dontrun{
a <- readSource("EUROSTAT")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Johanna Hoppe
}
