% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCAM.R
\name{readGCAM}
\alias{readGCAM}
\title{Read GCAM transportation data.}
\usage{
readGCAM(
  subtype = c("energyIntensity", "loadFactor", "histESdemand", "speedMotorized",
    "speedNonMotorized", "valueOfTimeMultiplier")
)
}
\arguments{
\item{subtype}{One of the possible subtypes, see default argument.}
}
\value{
magpie object
}
\description{
Read GCAM transportation data.
}
\examples{
\dontrun{
a <- readSource("GCAM", subtype = "histEDsemand")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Johnna Hoppe, Alois Dirnaichner
}
