% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMINDinputForTransportStandalone.R
\name{readREMINDinputForTransportStandalone}
\alias{readREMINDinputForTransportStandalone}
\title{Load information from a previous REMIND run}
\usage{
readREMINDinputForTransportStandalone(subtype)
}
\arguments{
\item{subtype}{Either "fuelCosts" or "esDemand"}
}
\description{
Load information from a previous REMIND run
}
\author{
Falk Benke
}
