% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustAnnualMileage.R
\name{toolAdjustAnnualMileage}
\alias{toolAdjustAnnualMileage}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustAnnualMileage(dt, completeData, filter, ariadneAdjustments = TRUE)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{completeData}{All combinations of region, period, univocalName and technology in EDGE-T decision tree}

\item{filter}{list of filters for specific branches in the upper decision tree, containing all associated
univocalNames}

\item{ariadneAdjustments}{switch on and off adjustments according to ARIADNE model intercomparison in 2022}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
