% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustCAPEXother.R
\name{toolAdjustCAPEXother}
\alias{toolAdjustCAPEXother}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustCAPEXother(dt, ISOcountries, yrs, completeData, GDPpcMER, filter)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{ISOcountries}{list of iso countries}

\item{yrs}{temporal resolution of EDGE-T model}

\item{completeData}{All combinations of region, period, univocalName and technology in EDGE-T decision tree}

\item{GDPpcMER}{GDP per capita based on market exchange rate}

\item{filter}{list of filters for specific branches in the upper decision tree, containing all associated univocalNames}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
