% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustEsDemand.R
\name{toolAdjustEsDemand}
\alias{toolAdjustEsDemand}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustEsDemand(dt, mapIso2region, completeData, filter)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{mapIso2region}{map iso countries to regions}

\item{completeData}{All combinations of region, period, univocalName and technology in EDGE-T decision tree}

\item{filter}{list of filters for specific branches in the upper decision tree, containing all associated
univocalNames}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
