% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustNonFuelOPEXtrackedFleet.R
\name{toolAdjustNonFuelOPEXtrackedFleet}
\alias{toolAdjustNonFuelOPEXtrackedFleet}
\title{Perform parameter specific adjustments on the input data}
\usage{
toolAdjustNonFuelOPEXtrackedFleet(dt, yrs, completeData)
}
\arguments{
\item{dt}{calculated raw data without adjustments}

\item{yrs}{temporal resolution of EDGE-T model}

\item{completeData}{complete EDGE-T decision tree}
}
\value{
a quitte object
}
\description{
Perform parameter specific adjustments on the input data
}
\author{
Johanna Hoppe
}
