% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolIEAharmonization.R
\name{toolIEAharmonization}
\alias{toolIEAharmonization}
\title{Harmonize the energy intensities to match the IEA energy balances regarding final energy.}
\usage{
toolIEAharmonization(enIntensity)
}
\arguments{
\item{enIntensity}{energy intensity}
}
\description{
We provide energy service trajectories. IEA energy balances have to be met and are \emph{not}
consistent with GCAM intensities and energy service trajectories.
Therefore we have to adjust energy intensities.
}
