% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareEurostat.R
\name{toolPrepareEurostat}
\alias{toolPrepareEurostat}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareEurostat(x)
}
\arguments{
\item{x}{the input data read via readSource, a magpie object}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories. Apply the full structure of the decision tree.
}
\author{
Johanna Hoppe
}
