% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrepareGCAM.R
\name{toolPrepareGCAM}
\alias{toolPrepareGCAM}
\title{Perform source specific transformations to ensure a compatible structure.}
\usage{
toolPrepareGCAM(magpieobj, subtype)
}
\arguments{
\item{magpieobj}{the input data read via readSource, a magpie object}

\item{sourcetype}{one of the different EDGE-T inputdata sources}
}
\value{
a quitte object
}
\description{
Map the source categories to the EDGE-T categories. Apply the full logit structure.
}
\author{
Johanna Hoppe, Alois Dirnaichner
}
